pyproject.toml
[project]
name = "physrisk-api"
version = "v0.30.0"
description = "OS-Climate Physical Risk API"
authors = [
  { name = "Joe Moorhouse", email = "5102656+joemoorhouse@users.noreply.github.com" },
]
requires-python = ">=3.8"
readme = "README.md"
license = { text = "Apache-2.0" }
keywords = ["Climate", "Finance"]
classifiers = [
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: MacOS",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: Unix",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.8",
  "Topic :: Office/Business :: Financial",
  "Topic :: Scientific/Engineering",
  "Topic :: Software Development",
]
dependencies = [
  "uvicorn",
  "fastapi",
  "python-dotenv",
  "fastapi-jwt-auth",
  "physrisk-lib>=0.37.0"
]

[project.urls]
Homepage = "https://github.com/os-climate/physrisk-api"
Repository = "https://github.com/os-climate/physrisk-api"
Downloads = "https://github.com/os-climate/physrisk-api/releases"
"Bug Tracker" = "https://github.com/os-climate/physrisk-api/issues"
Documentation = "https://github.com/os-climate/physrisk-api/tree/main/docs"
"Source Code" = "https://github.com/os-climate/physrisk-api"

[build-system]
requires = [
    "setuptools>=42",
    "wheel"
]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = "src/test"
addopts = "-v"

[tool.mypy]
warn_unreachable = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "click"
ignore_missing_imports = true

[tool.black]
line-length = 120
exclude = '''
/(
    \.git
  | \.tox
  | \venv
  | \.venv
  | \*.env
  | \build
  | \dist
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
profile = "black"

[tool.coverage.run]
omit = ["src/test/*"]
relative_files = true

